Update patch set 1
Patch Set 1:

I went through the whole init again to confirm existing paths etc. This isn't just a kang of Oppo scripts, but they are a really good model for future work.

Basic functionality is confirmed not broken (gsm, wifi, bluetooth) but with such a large change we should check everything.

I kept a lot of the original permissions as found in the Sony init, so if something was changed specifically for CM in the past it may need changing again.

Patch-set: 1