Update patch set 2
Patch Set 2:

Sorry for the wrong branch, and thanks for re-submitting.

The critical thing for fixing CYAN-4682 is making sure the sensors service is running after decryption.  The current mainline code doesn't do this, because the service gets killed as the temporary framework instance goes down after accepting the passphrase, and nobody starts it again.  The essence of the fix could probably be a one-liner, but the way I did it feels a bit cleaner and takes care of similar issues while I'm at it.  I'm happy to split that up into individual commits if needed.

I didn't move the service to late_start, so that the device could be rotated while on the passphrase screen (I am fairly sure this is necessary to achieve that, but I didn't test.)  Maybe that's not critical, but personally I'd prefer the option of typing the passphrase in landscape mode on my phone.

I moved the /data setup to post-fs-data, where (if I'm reading init.c and init.rc correctly) it will be run before an enabled "main" service starts in all cases.  That is, on an unencrypted /data, on the tmpfs mount, and on a decrypted /data.  I enabled the service mostly because it feels cleaner, and with the /data stuff in post-fs-data, the files should be set up before the service starts.  I'm also considering that, if not disabled, it will be restarted by init if it crashes, for whatever that's worth.

Will post-fs-data always run before the "main" class is started, or am I wrong about that?  Either way, I guess I don't mind making the service "disabled" and manually starting it from post-fs-data, if that's preferred.

By the way, there's an insmod /system/lib/modules/adsprpc.ko nearby; is that needed?  All the kernel configs for rhine-common appear to have this compiled in, not a module.

Patch-set: 2