Update patch set 3
Patch Set 3:

(6 comments)

I think I'll leave the Sony TA stuff as-is, because a) it's hard to preserve the dependencies and ordering between closed source stuff, b) most of it doesn't really need to persist, so it's OK if it goes to tmpfs and gets thrown away.

The exception is WLAN and Bluetooth MAC address provisioning.  It looks like the drivers will be loaded late in boot, after the TA stuff written to /data/etc/ is gone (in the case of encrypted device).  It seems there are default MAC addresses that are hard-coded and will be used if the real ones weren't found.  Luckily, most users avoid this because the initial unencrypted boot wrote that stuff to /data/etc/ and then got encrypted in place.  I did the wipe method (something like vdc cryptfs enablecrypto wipe <passphrase>) -- I believe this is not supported in the UI.  If it only affects users in unsupported configurations, and having a default MAC is not fatal to many users, is this not worth trying to fix?  If they have to, they can run taimport manually after encrypted boot.

The line review comments are documenting what I know about references to /data before post-fs-data, in case it's useful later.

Patch-set: 3