Update patch set 3
Patch Set 3:

Being able to rotate before entering the passphrase sounds okay, should be kept in, no reason to break that.

post-fs-data runs twice, so I guess your way is fine, dropping the disabled and keeping it class main.

The insmod should be removed, can be done in another change.



The WLAN and Bluetooth MAC is a niche use-case, so I'm fine with leaving it as it is.

Patch-set: 3