Update patch set 1
Patch Set 1:

It was actually bluetooth with the problem, as the hciattach service runs unpriveleged.  Wifi's mac-update runs as root in post-fs-data -- this should probably be fixed by moving the mac-update service to late_start, running as wifi:wifi, and removing the mac-update invocation in post-fs-data.

Here's a new change with 0640 and ownership change.  Both MAC addresses report as expected in About -> Status.  Now:

# adb shell ls -la /data/etc
-rw-r----- bluetooth bluetooth       17 1970-03-06 15:30 bluetooth_bdaddr
-rw-r----- wifi     wifi           17 1970-03-06 15:30 wlan_macaddr0
...

Btw, we both have 1970's as our modtime dates.  Kind of annoying, but I guess mostly harmless.  Is the clock losing time between boots?

Patch-set: 1