Update patch set 1
Patch Set 1:

With this, we could move this to sony/qcom-common, and name it custombootimg-dt.mk, only differences would be v1 vs v2.

What I have in mind is:
device/sony/foo/custombootimg.mk specifies v1 or vs2 (or just DTBTOOL_FLAGS := --force-v2 if v2), and then includes device/sony/qcom-common/custombootimg-dt.mk, which would be a copy of this but with the option of allowing to specifying dtb version.


There are some 8226 devices being brought up, they would benefit from this too.

Patch-set: 1