Update patch set 4
Patch Set 4:

Besides not being sure if all boards need it, the other reason I didn't put the flag in qcom-common, is that it would probably break the build of boards that have a custom ril.h that doesn't have the new struct member yet.

If I'm wrong about the build breaking, or it's still best to do that, PS2 in the qcom-common patch defines the flag.

Patch-set: 4