Update patch set 3
Patch Set 3:

Since installing this, I've had the screen get "stuck" off during a call, twice.  It was blank, not sure if backlight was on or if touches were received.  I had to hold power button, or connect with adb to reboot.  I couldn't get good logs, so I'm not sure if this is the cause.

Patch-set: 3