Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

shinano is still using the stock blobs.

I'm unable to test data right now, but if that doesn't work and we have to switch to AOSp blobs for that too, this commit would need some changes.

Either way, msm8974 has libril and the define which won't be needed for this, and should be moved or removed accordingly.

Patch-set: 1